Move type utility classes to type-utils.h (NFC)
This renames type-updating.h to type-utils.h and moves `TypeSeeker`
class, which is currently only used in wasm.cpp, to that file too. I
plan to use this elsewhere in #2451 and I think making this a separate
PR makes things easier to review.
